requirements.txt
flask==2.2.2
PyPDF2==2.11.1
python-docx==0.8.11
